Make the extration of URLs from Markdown more accurate
In source code files like Markdown or Python, URLs are often enclosed by
characters such as `''` or `()`. These characters are also valid within URLs.

This change looks for URLs that are enclosed by a pair of these characters,
and will ingore them when extracting the URLs. For example: in the markdown
`[link](https://example.com/abc)`, the trailing `)`, could be valid. But,
due to the presence of a preceeding `(`, it's very unlikely that it should be
included as part of the extracted URL.

This isn't a perfect solution, and could incorrectly remove a trailing `)`.
But, this is a pragmatic choice, which makes this code much more useful for
extracting URLs from source code files.